+++ packages/go_router/example/lib/named_routes.dart
@@ -59,7 +59,7 @@ void main() => runApp(App());
 /// The main app.
 class App extends StatelessWidget {
   /// Creates an [App].
-  App({Key? key}) : super(key: key);
+  App({super.key});
 
   /// The title of the app.
   static const String title = 'GoRouter Example: Named Routes';
@@ -105,7 +105,7 @@ class App extends StatelessWidget {
 /// The home screen that shows a list of families.
 class HomeScreen extends StatelessWidget {
   /// Creates a [HomeScreen].
-  const HomeScreen({Key? key}) : super(key: key);
+  const HomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -130,7 +130,7 @@ class HomeScreen extends StatelessWidget {
 /// The screen that shows a list of persons in a family.
 class FamilyScreen extends StatelessWidget {
   /// Creates a [FamilyScreen].
-  const FamilyScreen({required this.fid, Key? key}) : super(key: key);
+  const FamilyScreen({required this.fid, super.key});
 
   /// The id family to display.
   final String fid;
@@ -160,8 +160,7 @@ class FamilyScreen extends StatelessWidget {
 /// The person screen.
 class PersonScreen extends StatelessWidget {
   /// Creates a [PersonScreen].
-  const PersonScreen({required this.fid, required this.pid, Key? key})
-      : super(key: key);
+  const PersonScreen({required this.fid, required this.pid, super.key});
 
   /// The id of family this person belong to.
   final String fid;