+++ packages/camera/camera/CHANGELOG.md
@@ -1,5 +1,5 @@
-
 ## 0.10.4
+
 * Adds NV21 as an image streaming option for Android.
 
 ## 0.10.3