+++ packages/camera/camera/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Adds NV21 as an image streaming option for Android.
 
+## 0.10.3+1
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
+
 ## 0.10.3
 
 * Adds back use of Optional type.